Split local data provider.
