+++ data/lib/data.dart
@@ -1,5 +1,5 @@
 library data;
 
 export 'di/data_di.dart';
-export 'providers/local_data_provider.dart';
+export 'providers/local_data_provider_impl.dart';
 export 'repositories/repositories.dart';
