+++ data/lib/providers/local_data_provider_impl.dart
@@ -1,0 +1,33 @@
+import 'package:data/providers/local_data_provider.dart';
+
+class LocalDataProviderImpl implements LocalDataProvider {
+  @override
+  Future<bool> contains(String key) {
+    // TODO: implement contains
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> delete(String key) {
+    // TODO: implement delete
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> deleteAll() {
+    // TODO: implement deleteAll
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<String?> read(String key) {
+    // TODO: implement read
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> write({required String key, required String value}) {
+    // TODO: implement write
+    throw UnimplementedError();
+  }
+}
